refactor: move logic into service dir
