+++ src/index.ts
@@ -1,7 +1,7 @@
-export * from "./profile/character";
-export * from "./profile/achievements";
-export * from "./profile/classjob";
-export * from "./freecompany/freecompany";
-export * from "./freecompany/members";
-export * from "./search/character-search";
-export * from "./search/freecompany-search";
+export * from "./service/profile/character";
+export * from "./service/profile/achievements";
+export * from "./service/profile/classjob";
+export * from "./service/freecompany/freecompany";
+export * from "./service/freecompany/members";
+export * from "./service/search/character-search";
+export * from "./service/search/freecompany-search";
