+++ src/service/core/paginated-page-parser.ts
