+++ src/service/freecompany/freecompany.ts
@@ -1,6 +1,6 @@
 import { Request } from "express";
 import { PageParser } from "../core/page-parser";
-import * as freecompany from "../lib/lodestone-css-selectors/freecompany/freecompany.json";
+import * as freecompany from "../../lib/lodestone-css-selectors/freecompany/freecompany.json";
 import { CssSelectorRegistry } from "../core/css-selector-registry";
 
 export class FreeCompany extends PageParser {
