+++ src/service/profile/character.ts
@@ -1,8 +1,8 @@
 import { Request } from "express";
 import { PageParser } from "../core/page-parser";
-import * as character from "../lib/lodestone-css-selectors/profile/character.json";
-import * as attributes from "../lib/lodestone-css-selectors/profile/attributes.json";
-import * as gearset from "../lib/lodestone-css-selectors/profile/gearset.json";
+import * as character from "../../lib/lodestone-css-selectors/profile/character.json";
+import * as attributes from "../../lib/lodestone-css-selectors/profile/attributes.json";
+import * as gearset from "../../lib/lodestone-css-selectors/profile/gearset.json";
 import { CssSelectorRegistry } from "../core/css-selector-registry";
 
 export class Character extends PageParser {
