+++ src/service/profile/classjob.ts
@@ -1,23 +1,24 @@
 import { Request } from "express";
 import { CssSelectorRegistry } from "../core/css-selector-registry";
-import * as members from "../lib/lodestone-css-selectors/freecompany/members.json";
+import * as classjob from "../../lib/lodestone-css-selectors/profile/classjob.json";
 import { PaginatedPageParser } from "../core/paginated-page-parser";
 
-export class FCMembers extends PaginatedPageParser {
+export class ClassJob extends PaginatedPageParser {
   protected getCSSSelectors(): CssSelectorRegistry {
-    return members;
+    return classjob;
   }
 
   protected getBaseURL(req: Request): string {
     return (
-      "https://na.finalfantasyxiv.com/lodestone/freecompany/" +
-      req.params.fcId +
-      "/member"
+      "https://na.finalfantasyxiv.com/lodestone/character/" +
+      req.params.characterId +
+      "/class_job"
     );
   }
 
   async parse(req: Request, columnsPrefix: string = ""): Promise<Object> {
     const fromSuper: any = await super.parse(req, columnsPrefix);
+    delete fromSuper.Pagination;
     return fromSuper;
   }
 }
